Refactor frontend scripts into modular structure
