+++ static/js/constants.js
@@ -1,0 +1,10 @@
+export const REQUIRED_FIELDS = [
+  { key: 'sector' },
+  { key: 'zacks', type: 'number' },
+  { key: 'tipranks', type: 'number' },
+  { key: 'sector_growth' }
+];
+
+export const CLASS_SUCCESS = 'status-success';
+export const CLASS_ERROR = 'status-error';
+export const FETCH_DELAY_MS = 300;
