+++ templates/index.html
@@ -165,6 +165,6 @@
           <div class="status error">⚠️ Sector Growth failed</div>
         {% endif %}
     </form>
-    <script src="{{ url_for('static', filename='save.js') }}"></script>
+    <script type="module" src="{{ url_for('static', filename='js/main.js') }}"></script>
 </body>
 </html>
